valid moves update and evaluation pseduo code
wrote a description for how the evaluation function will determine the
advantage or disadvantage of the computer (black team) based on number
of pieces and immediate capture moves available. yet to be implemented
in code

also fixed a problem with isValidMove, checks for the end point to be
empty instead of checking to see if it's a capture move.